build.xml recreates Makefile.
This build.xml uses the commons-net Apache library to re-create the
functionality of the included Makefile -- in large part, due to wholesale
copying of the commands from the Makefile, and the execution of a bash shell.

As a result, this will only work on a system with a bash shell.

In the future, I'm going to write custom Ant tasks to do this piping and echoing for me, in the
right way.